backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
httpx==0.25.2
tiktoken==0.5.2
python-dotenv==1.0.0
pydantic==2.5.1
